util/mime: use uint8_t for cnt
cnt should not exceed B64_BLOCK which is 4, so, it is unnecessary to use
uint32_t for the variable.

+++ src/util-decode-mime.c
@@ -1203,7 +1203,7 @@ static uint32_t ProcessBase64Remainder(
         const uint8_t *buf, const uint32_t len, MimeDecParseState *state, int force)
 {
     uint32_t buf_consumed = 0; /* consumed bytes from 'buf' */
-    uint32_t cnt = 0;
+    uint8_t cnt = 0;
     uint8_t block[B64_BLOCK];
 
     SCLogDebug("len %u force %d", len, force);
